Sync database system correction

+++ BoVeloManager/tools/Database/DatabaseClassInterface/DCI_kitamplate.cs
@@ -49,7 +49,7 @@ namespace BoVeloManager.tools {
         }
 
         public static int updateKitTemplate(KitTemplate kt) {
-            string q = DatabaseQuery.updateKitTemplate(kt.getId(), kt.getName(), kt.getCategory(), kt.getPrice(), kt.getProperties(), kt.getBikeQtt());
+            string q = DatabaseQuery.updateKitTemplate(kt.getId(), kt.getName(), (int)kt.getCategory(), kt.getPrice(), kt.getProperties(), kt.getBikeQtt());
             return Database.setData(q);
         }
     }
